Begin running routes to controllers through channels.
This is still a work in progress, but ultimately a controller function
will take an input channel and return an output channel. On each route
invocation, we can put the route params into the input channel of the
desired controller, and a template-context pair will appear in its
output channel. From there, we can just execute the template, providing
the context as an argument. Nice and easy.

Templates don't do any async work, but controllers can; so controllers
are channel based while templates are not.