Placeholder video view. Dependent on a local enfocus fix.
See issue #4: event.preventDefault is not preventing link navigation.
The temporary workaround was to check out a fork of enfocus, do a dirty
fix for the preventDefault issue, and symlink to it in the checkouts
directory, which leiningen uses to resolve dependencies for exactly this
kind of issue.

In the long run, we'll need an updated Enfocus.

WIP: Placeholder video view.