Basic Youtube integration in video view.
The main video view now embeds and autoplays the Youtube video. As soon
as the API code is dynamically loaded as a script fragment, we embed the
video player; as soon as the player is ready, we load the video
specified in the template context.

The new raiseyourgame.lib.youtube namespace contains functions which
control the sole player instance. Many details remain to be determined:
we need to ensure that only one player instance exists at a time, and we
probably want to set up something less rickety to handle all the
readiness data.