Only handle timecodes if playing the video that was loaded.
After a youtube/load call, there are several ways the player might start
playing a different video. The user might click an ad, or click a
related video after the current one ends. In those cases, we don't want
to try to prevent it -- that's Youtube's money right there. We just want
to turn off timecode handling.

Implemented by storing the youtube/load argument in the loaded-video
atom, and checking its id against the currently playing video when the
state changes to :playing. Updated video->id to url->id and gave it a
few more regexes to work with.