Add and test user and video creation API routes.
Also removed a level of context* from the routes definition, which does
make this diff harder to read. I realized that something like `(context*
"/users"` was not semantically useful, since `"/api/users/:id/videos"`
is technically a videos route, not a users route. Therefore, the
contexts were forming groupings that reflected the route grammar, not
the conceptual structure.