Sort user list by username; test; expand test helpers.
To implement sort orders, we added more queries to users.sql. The
user/get-users model function delegates to the correct query as needed.
This workaround is necessary because ORDER BY and ASC/DESC cannot be
dynamically altered in a PreparedStatement, but it's fine.

To test sort orders, we moved more functions up to test.helpers. In
particular, we added helpers/pg-collation, which helps us imitate
Postgres's "natural-english" sorting behavior.