Only accept user/pass as login credentials.
Previously, we accepted either email or username, but we weren't testing
for it, and I don't want to deal with the situation where a user enters
an email address _as_ a username, enters someone else's email address as
a username, enters something like "jen@socal" as a username (confusing
a naive email regex), etc.

A lot of sites just have username login. It'll be fine.